% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Replacement Application}
\usage{
run_app(
  bgmaptype = "streets",
  mapboxkey = NULL,
  pghost = "localhost",
  pguser = "shiny",
  pgpass = NULL,
  pgdbname = "spatsampdata",
  filepath = "susomaps",
  susomapcrs = 3857,
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  ...
)
}
\arguments{
\item{bgmaptype}{the type of mapbox map, one of: "dark", "light", "outdoors",
"streets", "satellite", "satellite-streets".}

\item{mapboxkey}{mapbox key required for your basemap.}

\item{pghost}{Postgres host address, default is localhost}

\item{pguser}{Postgres user.}

\item{pgpass}{Postgres password.}

\item{pgdbname}{Postgress database name.}

\item{filepath}{Name of the directory where the files for the background maps are stored during processing. If
this folder does not exists, it will be created in the current working directory.}

\item{susomapcrs}{Coordinate reference system for fieldwork maps. Must be compatible with Survey Solutions. If you are
not sure, leave it at the default 3857, web-mercator.}

\item{onStart}{A function that will be called before the app is actually run.
This is only needed for \code{shinyAppObj}, since in the \code{shinyAppDir}
case, a \code{global.R} file can be used for this purpose.}

\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}

\item{enableBookmarking}{Can be one of \code{"url"}, \code{"server"}, or
\code{"disable"}. The default value, \code{NULL}, will respect the setting from
any previous calls to  \code{\link[shiny:enableBookmarking]{enableBookmarking()}}. See \code{\link[shiny:enableBookmarking]{enableBookmarking()}}
for more information on bookmarking your app.}

\item{uiPattern}{A regular expression that will be applied to each \code{GET}
request to determine whether the \code{ui} should be used to handle the
request. Note that the entire request path must match the regular
expression in order for the match to be considered successful.}

\item{...}{arguments to pass to golem_opts.
See `?golem::get_golem_options` for more details.}
}
\description{
This will start the Grid Replacment Application which is part of the Survey Solutions Spatial
Sampling toolbox. To use it, make sure you have access to the frame files generated by the Spatial Sampling
application.
}
