% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_storage.R
\name{utils_storage}
\alias{utils_storage}
\title{Storage Backend Abstraction Layer}
\description{
Provides a unified, backend-agnostic API for reading and writing geospatial
data (vector layers and rasters).  All functions accept a `backend` list
produced by `get_storage_backend()` and dispatch to either the
PostgreSQL/PostGIS helpers in `utils_postgres.R` or to local-filesystem
helpers that read/write ESRI Shapefiles and GeoTIFF files.

**Directory layout for the local backend:**
```
<local_dir>/
  shapefiles/   # ESRI Shapefiles (.shp + sidecar files)
  rasterfiles/  # GeoTIFF rasters  (.tif)
  basemaps/     # Downloaded basemap GeoTIFFs (persistent cache)
```
}
\keyword{internal}
